Directly benchmark `GitRepo.checkout` instead of `Component.checkout`
