+++ tests/test_dependency_mgmt.py
@@ -9,6 +9,7 @@ import pytest
 import json
 from unittest.mock import patch
 from pathlib import Path
+from typing import Iterable
 
 from commodore import dependency_mgmt
 from commodore.config import Config
@@ -16,7 +17,26 @@ from commodore.component import Component
 from commodore.helpers import relsymlink
 from commodore.inventory import Inventory
 
-from bench_component import setup_components_upstream
+
+def setup_components_upstream(tmp_path: Path, components: Iterable[str]):
+    # Prepare minimum component directories
+    upstream = tmp_path / "upstream"
+    component_urls = {}
+    component_versions = {}
+    for component in components:
+        repo_path = upstream / component
+        component_urls[component] = f"file://#{repo_path.resolve()}"
+        component_versions[component] = None
+        repo = git.Repo.init(repo_path)
+
+        class_dir = repo_path / "class"
+        class_dir.mkdir(parents=True, exist_ok=True)
+        (class_dir / "defaults.yml").touch(exist_ok=True)
+
+        repo.index.add(["class/defaults.yml"])
+        repo.index.commit("component defaults")
+
+    return component_urls, component_versions
 
 
 @pytest.fixture
